requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
anthropic>=0.20.0
python-dotenv>=1.0.0
lxml>=4.9.0
